+++ src/GRA.Domain.Repository/IEnteredSchoolRepository.cs
@@ -6,7 +6,7 @@ namespace GRA.Domain.Repository
 {
     public interface IEnteredSchoolRepository : IRepository<Model.EnteredSchool>
     {
-        Task ConvertSchoolAsync(int userId, int enteredSchoolId, int schoolId);
+        Task ConvertSchoolAsync(EnteredSchool enteredSchool, int schoolId);
         Task<DataWithCount<ICollection<EnteredSchool>>> GetPaginatedListAsync(int siteId,
             int skip,
             int take);